Using new Play version to parse Double

+++ app/controllers/ViewEnrolment.scala
@@ -1,4 +1,4 @@
 package controllers
 import anorm._
 
-case class ViewEnrolment(id: Long, course: String,dni: String, note : Int)
+case class ViewEnrolment(id: Long, course: String,dni: String, note : Double)
